Working.

+++ cifar_reader.h
@@ -9,6 +9,6 @@ public:
   bool read_bin(std::string filepath);
   
   std::vector<int> labels_;
-  std::vector< std::vector<int> > images_;
+  std::vector< std::vector<float> > images_;
 };
